Move plugin build files -> Resources (#54).
The Plugin Build Files page is actually not only about plugins,
these files are used when buildingalso the core. Also,
the current location under Legacy Plugins makes no sense, it's
used in all sorts of plugin builds.

Move it to Resources. This place is where we collect various,
valuable links, and this is indeed such a case.